+++ CumulusMX/EcowittCloudStation.cs
@@ -616,6 +616,17 @@ namespace CumulusMX
 					DoPressure(ConvertUnits.PressMBToUser(slp), Utils.FromUnixTime(data.pressure.absolute.time));
 				}
 
+				// === LDS ===
+				try
+				{
+					ProcessLDS(data, thisStation);
+				}
+				catch (Exception ex)
+				{
+					cumulus.LogErrorMessage($"ProcessCurrentData: Error in LDS data - {ex.Message}");
+				}
+
+
 				thisStation.DoForecast("", false);
 
 				cumulus.BatteryLowAlarm.Triggered = batteryLow;
@@ -1132,6 +1143,120 @@ namespace CumulusMX
 				if (data.water_leak.leak_ch4 != null)
 				{
 					station.DoLeakSensor(data.water_leak.leak_ch4.value, 4);
+				}
+			}
+		}
+
+		private static void ProcessLDS(EcowittApi.CurrentDataData data, WeatherStation station)
+		{
+			if (data.ch_lds1 != null)
+			{
+				if (data.ch_lds1.air_ch1 != null)
+				{
+					decimal? dist = data.ch_lds1.air_ch1.unit switch
+					{
+						"mm" => ConvertUnits.LaserMmToUser(data.ch_lds1.air_ch1.value.Value),
+						"cm" => ConvertUnits.LaserMmToUser(data.ch_lds1.air_ch1.value.Value / 10),
+						"in" => ConvertUnits.LaserInchesToUser(data.ch_lds1.air_ch1.value.Value),
+						"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds1.air_ch1.value.Value / 12),
+						_ => data.ch_lds1.air_ch1.value
+					};
+					station.DoLaserDistance(dist, 1);
+				}
+				if (data.ch_lds1.depth_ch1 != null)
+				{
+					decimal? dist = data.ch_lds1.depth_ch1.unit switch
+					{
+						"mm" => ConvertUnits.LaserMmToUser(data.ch_lds1.depth_ch1.value.Value),
+						"cm" => ConvertUnits.LaserMmToUser(data.ch_lds1.depth_ch1.value.Value),
+						"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds1.depth_ch1.value.Value / 12),
+						_ => data.ch_lds1.depth_ch1.value
+					};
+					station.DoLaserDepth(dist, 1);
+				}
+
+				if (data.ch_lds2 != null)
+				{
+					if (data.ch_lds2.air_ch2 != null)
+					{
+						decimal? dist = data.ch_lds2.air_ch2.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds2.air_ch2.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds2.air_ch2.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds2.air_ch2.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds2.air_ch2.value.Value / 12),
+							_ => data.ch_lds2.air_ch2.value
+						};
+						station.DoLaserDistance(dist, 2);
+					}
+					if (data.ch_lds2.depth_ch2 != null)
+					{
+						decimal? dist = data.ch_lds2.depth_ch2.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds2.depth_ch2.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds2.depth_ch2.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds2.depth_ch2.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds2.depth_ch2.value.Value / 12),
+							_ => data.ch_lds2.depth_ch2.value
+						};
+						station.DoLaserDepth(dist, 1);
+					}
+				}
+
+				if (data.ch_lds3 != null)
+				{
+					if (data.ch_lds3.air_ch3 != null)
+					{
+						decimal? dist = data.ch_lds3.air_ch3.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds3.air_ch3.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds3.air_ch3.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds3.air_ch3.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds3.air_ch3.value.Value / 12),
+							_ => data.ch_lds3.air_ch3.value
+						};
+						station.DoLaserDistance(dist, 2);
+					}
+					if (data.ch_lds3.depth_ch3 != null)
+					{
+						decimal? dist = data.ch_lds3.depth_ch3.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds3.depth_ch3.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds3.depth_ch3.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds3.depth_ch3.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds3.depth_ch3.value.Value / 12),
+							_ => data.ch_lds3.depth_ch3.value
+						};
+						station.DoLaserDepth(dist, 1);
+					}
+				}
+
+				if (data.ch_lds4 != null)
+				{
+					if (data.ch_lds4.air_ch4 != null)
+					{
+						decimal? dist = data.ch_lds4.air_ch4.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds4.air_ch4.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds4.air_ch4.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds4.air_ch4.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds4.air_ch4.value.Value / 12),
+							_ => data.ch_lds4.air_ch4.value
+						};
+						station.DoLaserDistance(dist, 2);
+					}
+					if (data.ch_lds4.depth_ch4 != null)
+					{
+						decimal? dist = data.ch_lds4.depth_ch4.unit switch
+						{
+							"mm" => ConvertUnits.LaserMmToUser(data.ch_lds4.depth_ch4.value.Value),
+							"cm" => ConvertUnits.LaserMmToUser(data.ch_lds4.depth_ch4.value.Value / 10),
+							"in" => ConvertUnits.LaserInchesToUser(data.ch_lds4.depth_ch4.value.Value),
+							"ft" => ConvertUnits.LaserInchesToUser(data.ch_lds4.depth_ch4.value.Value / 12),
+							_ => data.ch_lds4.depth_ch4.value
+						};
+						station.DoLaserDepth(dist, 1);
+					}
 				}
 			}
 		}