+++ CumulusMX/WM918Station.cs
@@ -23,7 +23,7 @@ namespace CumulusMX
 		public WM918Station(Cumulus cumulus)
 			: base(cumulus)
 		{
-			cumulus.Manufacturer = Cumulus.OREGON;
+			cumulus.Manufacturer = Cumulus.StationManufacturer.OREGON;
 			// station supplies rain rate
 			calculaterainrate = false;