+++ CumulusMX/WMR200Station.cs
@@ -51,7 +51,7 @@ namespace CumulusMX
 
 		public WMR200Station(Cumulus cumulus) : base(cumulus)
 		{
-			cumulus.Manufacturer = Cumulus.OREGONUSB;
+			cumulus.Manufacturer = Cumulus.StationManufacturer.OREGONUSB;
 			var devicelist = DeviceList.Local;
 			var station = devicelist.GetHidDeviceOrNull(Vendorid, Productid);